Fix layout switch delay in fzf in Vimwiki notes directory
Vimwiki notes names are in Russian, so I switch keyboard layout to
Russian when opening fzf in the Vimwiki directory. There was a delay
between opening fzf and layout switching, which allowed to type several
letters in English before the switch happened. This commit removes the
delay or at least reduces it significantly.